fundi-workbench/backend/requirements.txt
fastapi
uvicorn[standard]
google-genai
pydantic-settings
python-multipart
httpx
